digital_twin_experimentation/requirements.txt
azure-identity
azure-keyvault-secrets
azure-storage-blob
azure-iot-device
azure-digitaltwins-core
azure-eventhub
azure-functions
azure-eventhub-checkpointstoreblob-aio

# TODO: Make an API for the following service and deploy it on azure apps, also if possible, make the whole process of scaling up and down automatic